vkd3d-utils: Refactor away all event creation code from vkd3d-utils.
Prepares to remove the entire thing. It's unnecessary.
