+++ demos/gears.c
@@ -222,7 +222,7 @@ static void cxg_wait_for_previous_frame(struct cx_gears *cxg)
     {
         hr = ID3D12Fence_SetEventOnCompletion(fence->fence, v, fence->event);
         assert(SUCCEEDED(hr));
-        demo_wait_event(fence->event, INFINITE);
+        demo_wait_event(fence->event);
     }
 
     cxg->rt_idx = demo_swapchain_get_current_back_buffer_index(cxg->swapchain);
