+++ demos/triangle.c
@@ -144,7 +144,7 @@ static void cxt_wait_for_previous_frame(struct cx_triangle *cxt)
     {
         hr = ID3D12Fence_SetEventOnCompletion(fence->fence, v, fence->event);
         assert(SUCCEEDED(hr));
-        demo_wait_event(fence->event, INFINITE);
+        demo_wait_event(fence->event);
     }
 
     cxt->frame_idx = demo_swapchain_get_current_back_buffer_index(cxt->swapchain);
